Split giant Lucene test
